requirements.txt
asgiref==3.4.1
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
Werkzeug==2.0.2
